Refactor plugin organization into separate files
