+++ src/modules/plugins/core/config.js
@@ -1,0 +1,19 @@
+/**
+ * Core configuration options
+ * @type {Object}
+ */
+const core = {
+  /**
+  * The radius in hexagons of the world
+  * @type {number}
+  */
+  gridRadius: 25,
+
+  /**
+  * The radius in pixels of a hexagon within the grid
+  * @type {number}
+  */
+  hexRadius: 12
+};
+
+export default core;
