+++ src/modules/plugins/core/systems/index.js
@@ -1,0 +1,9 @@
+import BackgroundRenderer from "./BackgroundRenderer";
+import GridRenderer from "./GridRenderer";
+
+const systems = [
+  new BackgroundRenderer(),
+  new GridRenderer()
+];
+
+export default systems;
