+++ src/modules/plugins/plants/config.js
@@ -1,0 +1,19 @@
+/**
+ * Plant configuration options
+ * @type {Object}
+ */
+const plants = {
+  /**
+   * The percentage of the grid that will be covered in vegetation
+   * @type {number}
+   */
+  vegetationRate: 0.35,
+
+  /**
+   * Energy contained in a plant
+   * @type {number}
+   */
+  plantEnergy: 12
+};
+
+export default plants;
