+++ src/modules/plugins/plants/systems/index.js
@@ -1,0 +1,9 @@
+import PlantGenerator from "./PlantGenerator";
+import PlantRenderer from "./PlantRenderer";
+
+const systems = [
+  new PlantGenerator(),
+  new PlantRenderer()
+];
+
+export default systems;
